backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
networkx==3.2.1
httpx==0.25.2
pydantic==2.5.2
python-multipart==0.0.6